Add power manipulation revision compatability support
Adds new control options to power and menu allowing for explicit operation of 4v and 3v supplies.
It is critical that 4v and 3v are not enabled simultaneously, or for as little together as possible.
Therefore, 4v manipulation will explicitly check and manipulate 3v for exclusion. 3v is necessary to keep the CPU alive.

Currently, the board power supplies, capacitors and configuration should allow for enough time that the handoff between 4v and 3v can be done exclusively, ie that one can be disabled safely and there should be sufficient residual power for a window to turn the other on.
This has not been tested, and will need to be verified on the prototype.