(temp commit) disable init functions that appear to affect startup adversely
This is a collection of notes to fix items that have been disabled temporarily that were preventing the board from starting, and in some cases appear to be restarting.
Each of these items will have to be addressed, and analyzed against this revision 2.0, in particular:
- sleep manager (sleepmgr_enter_sleep())
- reset_processor()
- button_init() and button_service()

Additionally:
- GSM_modem_init() 
-CAN_elm_init()
have been disabled while troubleshooting the power issues. We are unable to determine yet why the shift register seems to not be honoring the bits to turn on the 5v power.