GSM power stop improvement
With the proper Rev2.0 board and power supplies, we were able to verify that holding the power pin succeeds in turning off the SIMCOM GSM chip. Because the production units will use the 5320a model, the issues potentially observed with the 7100a are not withstanding. If the 7100a is ever used, this function will need to be retested.