Add button interrupt driven power state functionality
Add an interrupt routine for the hardware button on the board.
Routine should deal with the implications that the button also starts the 4v supply, and takes further action after verifying that it is the button that was pressed.
Unfortunately, at this time of writing, a possible hardware bug exists that is preventing consistent operation at the lower power mode (3v only); but this commit contains only changes related to button functionality, and should work independently of the issue with that actual power state.