Button adjustment and notes on 3v
Adjust timing sensitivity of button to allow quicker press for bringing up full power mode.
Notes added for why our 3v checker only disables 3v.
There's good reason for this, testing has shown that there's a bit of a bug, and that we should be manipulating 3v exactly where we mean to, as it is timing critical.