Changed time update routines
Now the GPS time sync routine only launches if the SIM chip is out of
/RESET (physical reset) which it should be if power is on. if the chip
is in reset the GPS is not running so the time canot be queried.
Changed the validity of the Clock to 3 days, afterwards it will cause an
time sync to happen.
Found a bug in the EEPROM default values to initialize the timezone with
an uninitialized variable. Just set the default to Nevada, for good
measure.